Add distribution support for varying applications of mobile entities received by mobility providers.
With the use of the NodeMappingDistribution module it is possible
to simply change the assignment of mobile agent to different
node vectors, thus allowing a portion of mobile agent to have
a different set of active applications. It is also possible to
change the underlying node type, thus allowing the same type of
mobile agent to use LTE and or 802.11 hardware.

For configuration example see rover/simulations/mucFreiNetdLTE2dMulticast/omnetpp.ini
or docu in NodeMappingDistribution class.