add global OMNeT++ config options to set traci cmd interface
CFGID_[VADERE|SUMO|FLOW]_HOST create new configuration
options accessible via the command line interface in CROWNET

The options are --vadere-host --sumo-host --flow-host

Example:
 ./CROWNET -c someConfig -i omnetpp.ini --vadere-host=vadere001:9999

This will override the host/port setup in the TraCI connection
between Vadere and OMNeT++
This is needed if multiple containerized simulations are create
programmatically. In this case the name of the host (i.e. Container
name) is not known in advance.