fix: allow non zero convBound in sumo networks and log all bounds.
Updates the handling of cooridnate transformations based on sumo networks.
Each sumo-net.xml has a <location> tag with an offset, convBound,
orignBound and projection attribute. The offset gives the offset after
applying the transformation, setting the lower left point to zero. The
convBound gives the bound of all points defined in the network. OMNeT
uses this bound as the bounding box to remove unnecessary spaces around
the network. As OMNeT uses the upper right coordinate as origin with,
positive y down, the convBound is needed for the transformation.