fix: communicate Rsd_id with beacon. Deriving rsd_id from current rsd is error prone.
Errors occur when a domain change (changing base stations) happens
between the measurement time and receiving time.
Error A: the measurement belongs to the old RSD and transmission and reception on the mac Layer
happens before the domain change occurs but the reception is delayed
until after the change due to processing delay in the receiver. Beacon A
belongs to the prior RSD in this case but would be matched to the new on.
Therefore a fix was introduced in the Entry class to check this mismatch
and use the old RSD!.

Error B: the measurement is before the domain change happens AND belong
to and new cell but the transmission
is delayed in the mac layer for the sender. No new time stamp is added to
the packet so this cannot be marked on the packet. The transmission is
delayed until the domain change happens, which makes the node a viable receiver.
Due to the fix introduced to mitigate error A, beacon B will now be associated
with the wrong RSD

solution: attach the RSD information into the beacon so the receiver can
correctly match each measurement to the correct RSD